fix(avp): improve config
