+++ cmd/avp/main.go
@@ -1,14 +1,9 @@
 package main
 
 import (
-	"fmt"
 	"net/http"
 	_ "net/http/pprof"
-	"os"
-	"path"
 
-	iavp "github.com/pion/ion-avp/pkg"
-	"github.com/pion/ion-avp/pkg/elements"
 	log "github.com/pion/ion-log"
 	conf "github.com/pion/ion/pkg/conf/avp"
 	"github.com/pion/ion/pkg/discovery"
@@ -20,22 +15,7 @@ func init() {
 	fixByFunc := []string{}
 	log.Init(conf.Avp.Log.Level, fixByFile, fixByFunc)
 
-	elems := make(map[string]iavp.ElementFun)
-	if conf.Element.Webmsaver.On {
-		if _, err := os.Stat(conf.Element.Webmsaver.Path); os.IsNotExist(err) {
-			if err = os.MkdirAll(conf.Element.Webmsaver.Path, 0755); err != nil {
-				log.Errorf("make dir error: %v", err)
-			}
-		}
-		elems["webmsaver"] = func(rid, pid, tid string, config []byte) iavp.Element {
-			filewriter := elements.NewFileWriter(path.Join(conf.Element.Webmsaver.Path, fmt.Sprintf("%s-%s.webm", rid, pid)))
-			webm := elements.NewWebmSaver()
-			webm.Attach(filewriter)
-			return webm
-		}
-	}
-
-	avp.InitAVP(conf.Avp, elems)
+	avp.InitAVP(conf.Avp)
 }
 
 func main() {
